A diy Large format CoreXY 3D printer, uses mostly 3D printed and Cnc cut parts, but can be adjusted to use all 3d printed parts
It was created as a school project and took about 3 months to get it all figured out. It is based on the E3D V6 hotend and the MKS Monster8 V2 motherboard. The frame is created using 2020 and 2040 Aluminum extrusions. The motion system uses MGN12 size linear rails and 8mm linear rods and bearings. The print area ir approximately 400x400x550mm.